+++ raiden-dapp/src/components/GeneralMenu.vue
@@ -1,36 +1,38 @@
 <template>
   <div class="general-screen-menu">
-    <v-row class="general-screen-menu__account-details" no-gutters>
-      <v-col cols="12">
-        <div class="general-screen-menu__account-details--title">
-          {{ $t('general-menu.account-details') }}
-        </div>
-      </v-col>
-    </v-row>
-    <v-row no-gutters>
-      <v-col cols="2">
-        <span class="general-screen-menu__account-details--address">
-          {{ $t('general-menu.address') }}
-        </span>
-      </v-col>
-      <v-col cols="10">
-        <span class="general-screen-menu__account-details--address">
-          <address-display :address="defaultAccount" full-address />
-        </span>
-      </v-col>
-    </v-row>
-    <v-row no-gutters>
-      <v-col cols="2">
-        <span class="general-screen-menu__account-details--eth">
-          {{ $t('general-menu.currency') }}
-        </span>
-      </v-col>
-      <v-col cols="10">
-        <span class="general-screen-menu__account-details--eth">
-          {{ accountBalance | decimals }}
-        </span>
-      </v-col>
-    </v-row>
+    <div v-if="!loading && defaultAccount">
+      <v-row class="general-screen-menu__account-details" no-gutters>
+        <v-col cols="12">
+          <div class="general-screen-menu__account-details--title">
+            {{ $t('general-menu.account-details') }}
+          </div>
+        </v-col>
+      </v-row>
+      <v-row no-gutters>
+        <v-col cols="2">
+          <span class="general-screen-menu__account-details--address">
+            {{ $t('general-menu.address') }}
+          </span>
+        </v-col>
+        <v-col cols="10">
+          <span class="general-screen-menu__account-details--address">
+            <address-display :address="defaultAccount" full-address />
+          </span>
+        </v-col>
+      </v-row>
+      <v-row class="general-screen-menu__account-details__eth" no-gutters>
+        <v-col cols="2">
+          <span class="general-screen-menu__account-details__eth--currency">
+            {{ $t('general-menu.currency') }}
+          </span>
+        </v-col>
+        <v-col cols="10">
+          <span class="general-screen-menu__account-details__eth--balance">
+            {{ accountBalance | decimals }}
+          </span>
+        </v-col>
+      </v-row>
+    </div>
     <v-list two-line class="general-screen-menu__menu">
       <v-list-item
         v-for="(menuItem, index) in menuItems"
@@ -61,6 +63,7 @@
 <script lang="ts">
 import { Component, Mixins } from 'vue-property-decorator';
 import { mapState } from 'vuex';
+import { getLogsFromStore } from '@/utils/logstore';
 import NavigationMixin from '@/mixins/navigation-mixin';
 import AddressDisplay from '@/components/AddressDisplay.vue';
 
@@ -69,7 +72,7 @@ import AddressDisplay from '@/components/AddressDisplay.vue';
     AddressDisplay
   },
   computed: {
-    ...mapState(['defaultAccount', 'accountBalance'])
+    ...mapState(['loading', 'defaultAccount', 'accountBalance'])
   }
 })
 export default class GeneralMenu extends Mixins(NavigationMixin) {
@@ -86,8 +89,36 @@ export default class GeneralMenu extends Mixins(NavigationMixin) {
         route: () => {
           this.navigateToBackupState();
         }
+      },
+      {
+        icon: 'bug.svg',
+        title: this.$t('general-menu.menu-items.report-bugs-title') as string,
+        subtitle: this.$t(
+          'general-menu.menu-items.report-bugs-subtitle'
+        ) as string,
+        route: () => {
+          this.downloadLogs();
+        }
       }
     ];
+  }
+
+  /* istanbul ignore next */
+  async downloadLogs() {
+    const [lastTime, content] = await getLogsFromStore();
+    const filename = `raiden_${new Date(lastTime).toISOString()}.log`;
+    const file = new File([content], filename, { type: 'text/plain' });
+    const url = URL.createObjectURL(file);
+    const el = document.createElement('a');
+    el.href = url;
+    el.download = filename;
+    el.style.display = 'none';
+    document.body.appendChild(el);
+    el.click();
+    setTimeout(() => {
+      URL.revokeObjectURL(url);
+      document.body.removeChild(el);
+    }, 0);
   }
 }
 </script>
@@ -112,14 +143,19 @@ export default class GeneralMenu extends Mixins(NavigationMixin) {
     }
 
     &__eth {
-      color: rgba($color-white, 0.7);
-      font-size: 14px;
+      margin-bottom: 66px;
+
+      &__currency,
+      &__balance {
+        color: rgba($color-white, 0.7);
+        font-size: 14px;
+      }
     }
   }
 
   &__menu {
     background-color: transparent;
-    margin-top: 96px;
+    margin-top: 30px;
 
     &__list-items {
       border: solid 2px $secondary-text-color;